refactor(oracle): migrate ospf to per-protocol plugins
