+++ tools/oracle/engine/backends/scapy/packets.py
@@ -98,7 +98,6 @@ _SCAPY_LAYER_BY_LAYER: dict[str, str] = {
     "ipv6_fragment": "IPv6ExtHdrFragment",
     "ipv6_hop_by_hop": "IPv6ExtHdrHopByHop",
     "ipv6_routing": "IPv6ExtHdrRouting",
-    "ospf": "OSPF_Hdr",
     "radiotap": "RadioTap",
     "raw": "Raw",
     "rip": "RIP",
@@ -451,41 +450,6 @@ _SUPPORTED_FIELDS_BY_LAYER: dict[str, set[str]] = {
         "segleft",
         "type",
     },
-    "ospf": {
-        # Oracle-neutral OSPFv2 common-header fields from specs/layers/ospf.yaml.
-        "version",
-        "type",
-        "packet_length",
-        "length",
-        "len",
-        "router_id",
-        "area_id",
-        "checksum",
-        "autype",
-        "authentication",
-        # Per-type OSPFv2 body fields (Hello / DD / LSR / LSU / LSAck). The
-        # reference layer types these; unknown packet types or an explicit raw
-        # body fall back to opaque OSPF bytes.
-        "network_mask",
-        "hello_interval",
-        "options",
-        "router_priority",
-        "router_dead_interval",
-        "designated_router",
-        "backup_designated_router",
-        "neighbors",
-        "interface_mtu",
-        "dd_flags",
-        "dd_sequence_number",
-        "lsa_headers",
-        "requests",
-        "lsas",
-        "num_lsas",
-        "body",
-        "body_hex",
-        "raw",
-        "raw_body",
-    },
     "radiotap": {
         "antenna",
         "channel_flags",
@@ -659,8 +623,6 @@ def _build_layer(plan: PacketPlan, stack: list[str], index: int, scapy_all: Any)
 
     if layer == "raw":
         return scapy_all.Raw(load=_payload_bytes(fields))
-    if layer == "ospf":
-        return _ospf(fields, stack, index, scapy_all)
     if layer == "ipv6_hop_by_hop":
         return _ipv6_hop_by_hop(fields, stack, index, scapy_all)
     if layer == "ipv6_destination_options":
@@ -699,240 +661,6 @@ def _build_layer(plan: PacketPlan, stack: list[str], index: int, scapy_all: Any)
         return _zigbee_aps(fields, stack, index, scapy_all)
 
     raise ValueError(f"unsupported Scapy materialization layer: {layer}")
-
-
-# Oracle-neutral OSPFv2 packet-type domain names (specs/layers/ospf.yaml) mapped
-# to the RFC 2328 type codes the OSPF_Hdr type field carries.
-_OSPF_PACKET_TYPES: dict[str, int] = {
-    "hello": 1,
-    "database_description": 2,
-    "database-description": 2,
-    "link_state_request": 3,
-    "link-state-request": 3,
-    "link_state_update": 4,
-    "link-state-update": 4,
-    "link_state_ack": 5,
-    "link-state-ack": 5,
-}
-# Oracle-neutral OSPF AuType domain names mapped to the AuType codes.
-_OSPF_AUTYPES: dict[str, int] = {
-    "null": 0,
-    "simple": 1,
-    "simple_password": 1,
-    "cryptographic": 2,
-}
-
-
-def _ospf(
-    fields: Mapping[str, JSONObject],
-    stack: Sequence[str],
-    index: int,
-    scapy_all: Any,
-) -> Any:
-    """Materialize an OSPFv2 packet for the ``[ipv4, ospf, payload]`` stack.
-
-    The oracle-neutral common-header field names from ``specs/layers/ospf.yaml``
-    map to the ``scapy.contrib.ospf`` ``OSPF_Hdr`` fields (version/type/len/src/
-    area/chksum/authtype/authdata); the per-type body fields map to the matching
-    ``OSPF_Hello``/``OSPF_DBDesc``/``OSPF_LSReq``/``OSPF_LSUpd``/``OSPF_LSAck``
-    bodies. Unknown packet types or an explicit raw ``body`` fall back to opaque
-    OSPF bytes after the common header so they round-trip without Scapy
-    re-interpreting them.
-    """
-
-    scapy_ospf = import_scapy()["ospf"]
-    ospf_fields = _layer_fields_for_stack_index(fields, stack, index)
-    packet_type = _ospf_packet_type(_required_field(ospf_fields, "ospf", "type"))
-    header_kwargs = _ospf_header_kwargs(ospf_fields, packet_type)
-    header = scapy_ospf.OSPF_Hdr(type=packet_type, **header_kwargs)
-
-    body = _ospf_raw_body(ospf_fields)
-    if body is not None:
-        return header / scapy_all.Raw(load=body)
-
-    if packet_type == 1:
-        return header / _ospf_hello(ospf_fields, scapy_ospf)
-    if packet_type == 2:
-        return header / _ospf_database_description(ospf_fields, scapy_ospf)
-    if packet_type == 3:
-        return header / _ospf_link_state_request(ospf_fields, scapy_ospf)
-    if packet_type == 4:
-        return header / _ospf_link_state_update(ospf_fields, scapy_ospf)
-    if packet_type == 5:
-        return header / _ospf_link_state_ack(ospf_fields, scapy_ospf)
-    return header
-
-
-def _ospf_header_kwargs(fields: Mapping[str, object], packet_type: int) -> dict[str, Any]:
-    kwargs: dict[str, Any] = {
-        "version": _int(_optional_field(fields, "version"), 2),
-    }
-    router_id = _optional_field(fields, "router_id")
-    if router_id is not None:
-        kwargs["src"] = _text(router_id, "0.0.0.0")
-    area_id = _optional_field(fields, "area_id")
-    if area_id is not None:
-        kwargs["area"] = _text(area_id, "0.0.0.0")
-    if _optional_field(fields, "packet_length", "length", "len") is not None:
-        kwargs["len"] = _int(_optional_field(fields, "packet_length", "length", "len"), 0)
-    if _optional_field(fields, "checksum") is not None:
-        kwargs["chksum"] = _int(_optional_field(fields, "checksum"), 0)
-    if _optional_field(fields, "autype") is not None:
-        kwargs["authtype"] = _ospf_autype(_optional_field(fields, "autype"))
-    authentication = _optional_field(fields, "authentication")
-    if authentication is not None:
-        kwargs["authdata"] = int.from_bytes(_bytes_field(authentication, pad_to=8)[:8], "big")
-    return kwargs
-
-
-def _ospf_packet_type(value: object) -> int:
-    if isinstance(value, str):
-        normalized = value.lower().replace("-", "_")
-        if normalized in _OSPF_PACKET_TYPES:
-            return _OSPF_PACKET_TYPES[normalized]
-        return int(normalized, 0)
-    return _int(value, 0)
-
-
-def _ospf_autype(value: object) -> int:
-    if isinstance(value, str):
-        normalized = value.lower().replace("-", "_")
-        if normalized in _OSPF_AUTYPES:
-            return _OSPF_AUTYPES[normalized]
-        return int(normalized, 0)
-    return _int(value, 0)
-
-
-def _ospf_raw_body(fields: Mapping[str, object]) -> bytes | None:
-    value = _optional_field(fields, "body", "body_hex", "raw_body", "raw")
-    if value is None:
-        return None
-    return _bytes_field(value)
-
-
-def _ospf_neighbors(value: object) -> list[str]:
-    if value is None:
-        return []
-    if isinstance(value, (str, bytes, bytearray)):
-        return [_text(value, "0.0.0.0")]
-    if isinstance(value, Sequence):
-        return [_text(item, "0.0.0.0") for item in value]
-    return []
-
-
-def _ospf_hello(fields: Mapping[str, object], scapy_ospf: Any) -> Any:
-    kwargs: dict[str, Any] = {
-        "neighbors": _ospf_neighbors(_optional_field(fields, "neighbors")),
-    }
-    if _optional_field(fields, "network_mask") is not None:
-        kwargs["mask"] = _text(_optional_field(fields, "network_mask"), "0.0.0.0")
-    if _optional_field(fields, "hello_interval") is not None:
-        kwargs["hellointerval"] = _int(_optional_field(fields, "hello_interval"), 0)
-    if _optional_field(fields, "options") is not None:
-        kwargs["options"] = _int(_optional_field(fields, "options"), 0)
-    if _optional_field(fields, "router_priority") is not None:
-        kwargs["prio"] = _int(_optional_field(fields, "router_priority"), 0)
-    if _optional_field(fields, "router_dead_interval") is not None:
-        kwargs["deadinterval"] = _int(_optional_field(fields, "router_dead_interval"), 0)
-    if _optional_field(fields, "designated_router") is not None:
-        kwargs["router"] = _text(_optional_field(fields, "designated_router"), "0.0.0.0")
-    if _optional_field(fields, "backup_designated_router") is not None:
-        kwargs["backup"] = _text(_optional_field(fields, "backup_designated_router"), "0.0.0.0")
-    return scapy_ospf.OSPF_Hello(**kwargs)
-
-
-def _ospf_database_description(fields: Mapping[str, object], scapy_ospf: Any) -> Any:
-    kwargs: dict[str, Any] = {
-        "lsaheaders": _ospf_lsa_headers(_optional_field(fields, "lsa_headers"), scapy_ospf),
-    }
-    if _optional_field(fields, "interface_mtu") is not None:
-        kwargs["mtu"] = _int(_optional_field(fields, "interface_mtu"), 0)
-    if _optional_field(fields, "options") is not None:
-        kwargs["options"] = _int(_optional_field(fields, "options"), 0)
-    if _optional_field(fields, "dd_flags") is not None:
-        kwargs["dbdescr"] = _int(_optional_field(fields, "dd_flags"), 0)
-    if _optional_field(fields, "dd_sequence_number") is not None:
-        kwargs["ddseq"] = _int(_optional_field(fields, "dd_sequence_number"), 0)
-    return scapy_ospf.OSPF_DBDesc(**kwargs)
-
-
-def _ospf_link_state_request(fields: Mapping[str, object], scapy_ospf: Any) -> Any:
-    requests = _optional_field(fields, "requests")
-    items: list[Any] = []
-    if isinstance(requests, Sequence) and not isinstance(requests, (str, bytes, bytearray)):
-        for entry in requests:
-            if not isinstance(entry, Mapping):
-                continue
-            items.append(
-                scapy_ospf.OSPF_LSReq_Item(
-                    type=_int(_optional_field(entry, "ls_type", "type"), 0),
-                    id=_text(_optional_field(entry, "link_state_id", "id"), "0.0.0.0"),
-                    adrouter=_text(
-                        _optional_field(entry, "advertising_router", "adrouter"),
-                        "0.0.0.0",
-                    ),
-                )
-            )
-    return scapy_ospf.OSPF_LSReq(requests=items)
-
-
-def _ospf_link_state_update(fields: Mapping[str, object], scapy_ospf: Any) -> Any:
-    kwargs: dict[str, Any] = {
-        "lsalist": _ospf_lsa_list(_optional_field(fields, "lsas"), scapy_ospf),
-    }
-    if _optional_field(fields, "num_lsas") is not None:
-        kwargs["lsacount"] = _int(_optional_field(fields, "num_lsas"), 0)
-    return scapy_ospf.OSPF_LSUpd(**kwargs)
-
-
-def _ospf_link_state_ack(fields: Mapping[str, object], scapy_ospf: Any) -> Any:
-    return scapy_ospf.OSPF_LSAck(
-        lsaheaders=_ospf_lsa_headers(_optional_field(fields, "lsa_headers"), scapy_ospf)
-    )
-
-
-def _ospf_lsa_headers(value: object, scapy_ospf: Any) -> list[Any]:
-    headers: list[Any] = []
-    if isinstance(value, Sequence) and not isinstance(value, (str, bytes, bytearray)):
-        for entry in value:
-            if not isinstance(entry, Mapping):
-                continue
-            headers.append(_ospf_lsa_header(entry, scapy_ospf))
-    return headers
-
-
-def _ospf_lsa_header(entry: Mapping[str, object], scapy_ospf: Any) -> Any:
-    kwargs: dict[str, Any] = {
-        "age": _int(_optional_field(entry, "ls_age", "age"), 0),
-        "options": _int(_optional_field(entry, "options"), 0),
-        "type": _int(_optional_field(entry, "ls_type", "type"), 0),
-        "id": _text(_optional_field(entry, "link_state_id", "id"), "0.0.0.0"),
-        "adrouter": _text(
-            _optional_field(entry, "advertising_router", "adrouter"),
-            "0.0.0.0",
-        ),
-        "seq": _int(_optional_field(entry, "ls_sequence_number", "seq"), 0x80000001),
-    }
-    if _optional_field(entry, "ls_checksum", "chksum") is not None:
-        kwargs["chksum"] = _int(_optional_field(entry, "ls_checksum", "chksum"), 0)
-    if _optional_field(entry, "length", "len") is not None:
-        kwargs["len"] = _int(_optional_field(entry, "length", "len"), 0)
-    return scapy_ospf.OSPF_LSA_Hdr(**kwargs)
-
-
-def _ospf_lsa_list(value: object, scapy_ospf: Any) -> list[Any]:
-    lsas: list[Any] = []
-    if isinstance(value, Sequence) and not isinstance(value, (str, bytes, bytearray)):
-        for entry in value:
-            if not isinstance(entry, Mapping):
-                continue
-            header = _ospf_lsa_header(entry, scapy_ospf)
-            body = _optional_field(entry, "body", "body_hex", "raw")
-            if body is not None:
-                lsas.append(header / import_scapy()["all"].Raw(load=_bytes_field(body)))
-            else:
-                lsas.append(header)
-    return lsas
 
 
 def _dns(fields: Mapping[str, JSONObject], scapy_all: Any) -> Any:
